requirements.txt
lark-oapi==1.5.3
python-dotenv==1.2.2
pydantic==2.12.5
structlog>=24.0.0

# dev
pytest>=8.0.0
pytest-asyncio>=0.24.0
